burn: change times when burn is possible to use burn windows
following feedback from 9999, we're doing a small modification to the burn spec in regards to when it can be invoked 🔥

The problem 9999 helped us identify:
Let's say the current nextBurnNounId is 900
If at noun id 900 there is no excess ETH (which is also the point of the mechanism), then at any moment from then when there is excess ETH, a burn can be invoked immediately, because the current noun id is > 900
We think this defeats the purpose of predictable burn times

The suggested fix:
There will be a burn window of a few nouns, e.g. 3, during which a burn can be invoked
If there was no burn in that window, then the next window will be at a multiple of nounIdsBetweenBurns

Example:
If we set the initial burn noun id to 1000, and the nounIdsBetweenBurns to 100, and the burn window to 3
The first burn can happen while the auctioned noun has id in the range 1000-1003.
Only one burn can happen in a window
Next burns can happen in the windows 1100-1103, 1200-1203, etc. regardless of whether a burn happened or not