add foundry profile with no optimizer
this can be used for local development. reduces compile times
significantly.

had to make this change in NounsDAOStorageV1Fork.sol because otherwise
we get a stack-too-deep error.
the stack-too-deep is because of the automatic getter that gets created

in the parallel NounsLogicDAOV2 this is not a public mapping, so no need
for it to be public here either