Fix for rewards contract following audit
This fix changes the way the proposal rewards calculate the auction revenue.
Instead of taking the user provided lastProposalId's creation timestamp,
we use the last *eligible* proposal that is <= lastProposalId.
The reason is we don't want users creating bogus proposals just for controlling
the end timestamp.

In addition, we also consider canceled proposals to be ineligible.
This is to prevent a user with quorum votes to create proposals, vote, and cancel,
just to get rewards.